[ALPS06295410] perf: print format right
Printing int64_t/uint64_t on Linux 64-bit or 32-bit platform is
architecture-dependent. For example, %ld for 64-bit int64_t and
%lld for 32-bit. Use a format specifier of its largest possible
type and expliicitly cast to it.

MTK-Commit-Id: 675fcb9a2ebb7f49e258c62b6d05a6783b684e94

CR-Id: ALPS06295410
Feature: [Module]Wi-Fi Driver
(cherry picked from commit 67f97f923e668ffb3a9516dea8984bb637039a1a)
(cherry picked from commit 512c0d2e2c4368491b33305a5c0ba1d4febc0971)
(cherry picked from commit 986b4711e37e4e0a67423e632da5b61802ccf6bb)

+++ os/linux/gl_kal.c
@@ -7991,9 +7991,14 @@ static uint32_t kalPerMonUpdate(IN struct ADAPTER *prAdapter)
 		if (txDiffBytes[i] < 0 || rxDiffBytes[i] < 0) {
 			/* overflow should not happen */
 			DBGLOG(SW4, WARN,
-				"[%d]wrong bytes: tx[%lu][%lu][%ld], rx[%lu][%lu][%ld],\n",
-				i, currentTxBytes, lastTxBytes, txDiffBytes[i],
-				currentRxBytes, lastRxBytes, rxDiffBytes[i]);
+				"[%d]wrong bytes: tx[%llu][%lld][%lld], rx[%llu][%lld][%lld],\n",
+				i,
+				(unsigned long long) currentTxBytes,
+				(long long) lastTxBytes,
+				(long long) txDiffBytes[i],
+				(unsigned long long) currentRxBytes,
+				(long long) lastRxBytes,
+				(long long) rxDiffBytes[i]);
 			goto fail;
 		}
